requirements.txt
# cryptography==43.0.1
# pyjwt==2.8.0
# flask==3.0.3

cryptography==43.0.1
pyjwt==2.8.0
